# poslappapi/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /roles:
    x-swagger-router-controller: rol.controller
    get:
      description: get the game system list
      operationId: getRoles
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRolListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new game system to the list
      operationId: addRol
      parameters:
        - in: body
          name: gamesystem
          description: The Game System to create.
          schema:
            $ref: "#/definitions/Roles"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetRolResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /roles/{id}:
    x-swagger-router-controller: rol.controller
    get:
      description: get a game system
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRolResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Roles:
    type: object
    properties:
      nombreRol:
        type: string
        description: Name of the Game System
    required:
      - nombreRol
  GetRolResponse:
    type: object
    properties:
      nombreRol:
        type: string
        description: Name of the Game System
      idRol:
        type: number
        description: An identifier
    required:
      - idRol
      - nombreRol
  GetRolListResponse:
    required:
      - roles
    properties:
      roles:
        type: array
        items:
          $ref: "#/definitions/GetRolResponse"
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string